wnsm_sync/requirements.txt
requests>=2.31.0
lxml>=4.9.3
python-dateutil>=2.8.2
paho-mqtt==1.6.1